drm/udl: Use drm mode config helpers
This changes the udl driver to use the drm mode config helpers for USB
suspend and resume. This allows us to remove the special support code
for this situation.

This also moves the USB suspend/resume functions next to the other USB
functions, which a more logical place to put them.

Signed-off-by: Scott Anderson <scott@anderso.nz>