drm/udl: Merge modesetting functions into udl_drv
Much of the modesetting functions are split across several smaller
files. It's cleaner and easier to manage to just have them all in the
same file.

This simply moves the contents of udl_connector.c and udl_modeset.c
into udl_drv.c, marks the previously extern functions as static, and
removes the prototypes from udl_drv.h.

No further cleanup of the code is done.

Signed-off-by: Scott Anderson <scott@anderso.nz>