drm/udl: Refactor initilization
Previously the initilization was spread out over several smaller
functions that were called once. This instead centralizes them all in
one place, when the device is probed, which makes it easier to follow
how the initlization works and leads to less code.

This also now makes use of the devm functions to simplify cleanup of
resources in error cases.

Signed-off-by: Scott Anderson <scott@anderso.nz>